requirements.txt
# Core CrewAI
crewai>=0.80.0
crewai-tools

# Data Science & ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.0.0

# Model Management
joblib>=1.3.0

# Dashboard
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# LLM Integration
openai>=1.0.0
langchain>=0.1.0
langchain-openai
